+++ src/mongocxx/include/mongocxx/v1/logger-fwd.hpp
@@ -27,6 +27,8 @@ class MONGOCXX_ABI_EXPORT logger;
 
 class default_logger;
 
+class MONGOCXX_ABI_EXPORT logger_guard;
+
 } // namespace v1
 } // namespace mongocxx